Refactor: move components files about product in product folder
