+++ week7/unique0902/src/components/Product/ProductForm/ProductForm.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Button from '../Button/Button';
+import Button from '../../Button/Button';
 import styles from './ProductForm.module.css';
 
 const ProductForm = ({ product, handleSubmit }) => {
