+++ week7/unique0902/src/components/Product/ProductRow/ProductRow.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useState } from 'react';
-import Button from '../Button/Button';
+import Button from '../../Button/Button';
 import EditProductForm from '../ProductForm/EditProductForm';
 import styles from './ProductRow.module.css';
